draw map scene to QImage
QImage is more optimized for drawing *to*. this is enough to improve drawing speed even while having to convert to and from a QPixmap!

of course, the ideal solution would be to move away from the graphics scene entirely to harness qwidget draw optimizations and skip the pixmap step, which we'll probably do soonish!