general codebase cleanup
optimized some imports and removed some unused redundant classes.

`EditDialog` seems to be a primitive version of what would eventually become `DatabaseSplitWidget`. `RpgSpinBox` is indistinguishable from `QSpinBox` and i struggle to see any need for customization. i'll add it again only if it's absolutely necessary, and only in places where a difference of functions is required.

i'm surprised that `RpgAudioLineEdit` is unused in the database, as there are multiple places where it'd fit perfectly. i'll probably genericize it in the near future and reintegrate it with the codebase.

`RpgSlider` survives the purge despite being empty because i've got plans for it (that being, i want my own step behavior), but that's out of scope for this commit.