defer flag reading to QSettings, part 1
this way we don't need a million setters and getters, at the cost of the available flags being a tad more opaque. given good enough documentation could be made for it (and the easyrpg wiki works for now), i think that's a fair sacrifice.

the configs aren't exactly being populated at the moment, though... i can't figure out how to do it and have it build. the official example just constructs a new QSettings instance every time it's needed...